// src/app/shared/model/movie-database.model.ts
export interface MovieDatabaseModel {
    id: number;
    category: string;
    date: string;
    movieId: number;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    status: string;
    userId: string;
    watched: boolean;
}